requirements.serving.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
appdirs==1.4.4
async-timeout==4.0.2
attrs==23.1.0
black==23.3.0
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
chumpy==0.70
click==8.1.3
contourpy==1.1.0
cycler==0.11.0
docker-pycreds==0.4.0
efficientnet-pytorch==0.7.1
fire==0.5.0
flake8==6.0.0
flake8-bugbear==23.6.5
flake8-comprehensions==3.13.0
fonttools==4.40.0
frozenlist==1.3.3
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.21.0
google-auth-oauthlib==1.0.0
grpcio==1.56.0
idna==3.4
imageio==2.31.1
importlib-metadata==6.7.0
importlib-resources==5.12.0
kiwisolver==1.4.4
lazy_loader==0.3
libcst==1.0.1
lightning-lite==1.8.0
lightning-utilities==0.3.0
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.7.1
mccabe==0.7.0
moreorless==0.4.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
numpy==1.23.1
oauthlib==3.2.2
opencv-python==4.8.0.74
packaging==23.1
pandas==2.0.3
pathspec==0.11.1
pathtools==0.1.2
Pillow==10.0.0
platformdirs==3.8.0
plotly==5.15.0
protobuf==4.23.3
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pyflakes==3.0.1
pyparsing==3.1.0
python-dateutil==2.8.2
pytorch-lightning==1.8.0
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.21.0
scipy
sentry-sdk==1.27.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
smplx==0.1.13
stdlibs==2022.10.9
tenacity==8.2.2
tensorboard==2.13.0
tensorboard-data-server==0.7.1
termcolor==2.3.0
tifffile==2023.7.4
toml==0.10.2
tomli==2.0.1
torchgeometry==0.1.2
torchmetrics==0.11.4
tqdm==4.65.0
trailrunner==1.4.0
trimesh==3.22.3
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
usort==1.0.7
wandb==0.15.4
Werkzeug==2.3.6
yarl==1.9.2
zipp==3.15.0
boto3
torchsummary
